requirements.txt
opencv-python==4.8.0.74
numpy==1.25.0
pandas==2.0.2
Pillow==9.5.0
seaborn==0.12.2
matplotlib==3.7.1
scikit-learn==1.2.2
streamlit==1.24.0